feat: move site config into root
