+++ components/site-header-mobile.tsx
@@ -9,7 +9,7 @@ import HeaderVector from "@/public/icons/menu-burger.svg"
 
 import { LangProps } from "@/types/common"
 import { NavItem } from "@/types/nav"
-import { siteConfig } from "@/config/site"
+import { siteConfig } from "@/site-config"
 import { cn } from "@/lib/utils"
 import { useAppSettings } from "@/hooks/useAppSettings"
 import {
